# model/TodoAppDotNet.openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TodoApp API",
        "version": "1.0.0",
        "description": "A REST API for managing TODO lists and tasks"
    },
    "paths": {
        "/users": {
            "get": {
                "description": "List all users",
                "operationId": "ListUsers",
                "parameters": [
                    {
                        "name": "maxResults",
                        "in": "query",
                        "description": "Pagination max results constraint",
                        "schema": {
                            "type": "number",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Pagination max results constraint"
                        }
                    },
                    {
                        "name": "nextToken",
                        "in": "query",
                        "description": "Pagination token constraint",
                        "schema": {
                            "type": "string",
                            "maxLength": 1000,
                            "minLength": 1,
                            "description": "Pagination token constraint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ListUsers 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListUsersResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ValidationException 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationExceptionResponseContent"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user",
                "operationId": "CreateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequestContent"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "CreateUser 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateUserResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ValidationException 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationExceptionResponseContent"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "ConflictException 409 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConflictExceptionResponseContent"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/lookup": {
            "get": {
                "description": "Get a user by their email address",
                "operationId": "GetUserByEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email with validation constraints",
                        "schema": {
                            "type": "string",
                            "maxLength": 254,
                            "minLength": 5,
                            "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                            "description": "Email with validation constraints"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetUserByEmail 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserByEmailResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ValidationException 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationExceptionResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ResourceNotFoundException 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceNotFoundExceptionResponseContent"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "delete": {
                "description": "Delete a user",
                "operationId": "DeleteUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "schema": {
                            "type": "string",
                            "description": "User identifier"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DeleteUser 200 response"
                    },
                    "400": {
                        "description": "ValidationException 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationExceptionResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ResourceNotFoundException 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceNotFoundExceptionResponseContent"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "description": "Get a user by ID",
                "operationId": "GetUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "schema": {
                            "type": "string",
                            "description": "User identifier"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetUser 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ValidationException 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationExceptionResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ResourceNotFoundException 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceNotFoundExceptionResponseContent"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a user",
                "operationId": "UpdateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequestContent"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User identifier",
                        "schema": {
                            "type": "string",
                            "description": "User identifier"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UpdateUser 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateUserResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ValidationException 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationExceptionResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ResourceNotFoundException 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceNotFoundExceptionResponseContent"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "ConflictException 409 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConflictExceptionResponseContent"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ConflictExceptionResponseContent": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "message"
                ]
            },
...
            "CreateUserRequestContent": {
                "type": "object",
                "description": "Structure for creating a new user",
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "description": "Username with validation constraints"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 5,
                        "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                        "description": "Email with validation constraints"
                    }
                },
                "required": [
                    "email",
                    "username"
                ]
            },
            "CreateUserResponseContent": {
                "type": "object",
                "description": "Structure for user response",
                "properties": {
                    "userId": {
                        "type": "string",
                        "description": "User identifier"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "description": "Username with validation constraints"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 5,
                        "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                        "description": "Email with validation constraints"
                    },
                    "createdAt": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "createdAt",
                    "email",
                    "userId",
                    "username"
                ]
            },
...
            "GetUserByEmailResponseContent": {
                "type": "object",
                "description": "Structure for user response",
                "properties": {
                    "userId": {
                        "type": "string",
                        "description": "User identifier"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "description": "Username with validation constraints"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 5,
                        "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                        "description": "Email with validation constraints"
                    },
                    "createdAt": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "createdAt",
                    "email",
                    "userId",
                    "username"
                ]
            },
            "GetUserResponseContent": {
                "type": "object",
                "description": "Structure for user response",
                "properties": {
                    "userId": {
                        "type": "string",
                        "description": "User identifier"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "description": "Username with validation constraints"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 5,
                        "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                        "description": "Email with validation constraints"
                    },
                    "createdAt": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "createdAt",
                    "email",
                    "userId",
                    "username"
                ]
            },
...
            "ListUsersResponseContent": {
                "type": "object",
                "properties": {
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserOutput"
                        }
                    },
                    "nextToken": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 1,
                        "description": "Pagination token constraint"
                    }
                },
                "required": [
                    "users"
                ]
            },
...
            "ResourceNotFoundExceptionResponseContent": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "message"
                ]
            },
...
            "UpdateUserRequestContent": {
                "type": "object",
                "description": "Structure for updating a user",
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "description": "Username with validation constraints"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 5,
                        "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                        "description": "Email with validation constraints"
                    }
                }
            },
            "UpdateUserResponseContent": {
                "type": "object",
                "description": "Structure for user response",
                "properties": {
                    "userId": {
                        "type": "string",
                        "description": "User identifier"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "description": "Username with validation constraints"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 5,
                        "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                        "description": "Email with validation constraints"
                    },
                    "createdAt": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "createdAt",
                    "email",
                    "userId",
                    "username"
                ]
            },
            "UserOutput": {
                "type": "object",
                "description": "Structure for user response",
                "properties": {
                    "userId": {
                        "type": "string",
                        "description": "User identifier"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "pattern": "^[a-zA-Z0-9_-]+$",
                        "description": "Username with validation constraints"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 254,
                        "minLength": 5,
                        "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
                        "description": "Email with validation constraints"
                    },
                    "createdAt": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "createdAt",
                    "email",
                    "userId",
                    "username"
                ]
            },
            "ValidationExceptionResponseContent": {
                "type": "object",
                "description": "Common error structures",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "message"
                ]
            }
        }
    },
    "servers": [
        {
            "url": "http://localhost:80",
            "description": "Development server"
        }
    ]
}